+++ src/js/card/removeFromFav.js
@@ -1,13 +1,10 @@
-import { ads } from '../helpers';
-import { getUserToken } from '../helpers/index';
 import { findCheckedCard } from './card';
 import { delFavItem } from '../favorites/remove-favorite';
 import { changeFavoriteStyle } from './card';
-import { requestUserFavorites } from '../helpers';
 
 // Замена иконки сердечко при удалении из Избранного
 export async function removeAddToFavorites(event) {
-  const selector = findCheckedCard(event);
+  const selector = findCheckedCard(event, 'card');
   await delFavItem(event);
   changeFavoriteStyle(
     selector,
@@ -16,57 +13,4 @@ export async function removeAddToFavorites(event) {
     'block',
     'none',
   );
-}
-
-// Проверка авторизованости пользователя, получение значения токена
-async function getToken() {
-  const userToken = await getUserToken();
-  if (userToken !== null) {
-    return userToken;
-  }
-  return false;
-}
-
-// Проверка авторизованости пользователя, получение true/false
-async function checkUserAuth() {
-  const userToken = await getUserToken();
-  if (userToken !== null) {
-    return true;
-  }
-  return false;
-}
-
-// Проверяем регистрацию юзера при загрузке страницы и вытягиваем id избранных карточек
-async function getAuthUserFavId() {
-  const userToken = await getToken();
-  if (userToken) {
-    return requestUserFavorites({ token: userToken })
-      .then(data => data.favourites)
-      .then(array => array.map(el => el._id))
-      .catch(error => console.log(error));
-  }
-  return false;
-}
-
-// находим сердечки из избранного и меняем цвет
-export async function checkUserFavIcons() {
-  const userToken = checkUserAuth();
-  const userFav = await getAuthUserFavId();
-  if (userToken !== true && userFav !== false) {
-    const selectors = Array.from(document.querySelectorAll('[data-id]'));
-    for (let item of userFav) {
-      selectors.map(card => {
-        if (card.dataset.id === item) {
-          changeFavoriteStyle(
-            card,
-            '.card__favorite-btn--orange',
-            '.card__favorite-btn',
-            'block',
-            'none',
-          );
-        }
-      });
-    }
-  }
-  return;
 }
